Move code to ISemanticFacts
