+++ src/Workspaces/SharedUtilitiesAndExtensions/Compiler/Core/Services/SemanticFacts/ISemanticFacts.cs
@@ -94,5 +94,10 @@ namespace Microsoft.CodeAnalysis.LanguageServices
 #nullable disable
 
         bool IsInsideNameOfExpression(SemanticModel semanticModel, SyntaxNode node, CancellationToken cancellationToken);
+
+        /// <summary>
+        /// Finds all local function definitions within the syntax references for a given <paramref name="symbol"/>
+        /// </summary>
+        ImmutableArray<IMethodSymbol> GetLocalFunctionSymbols(Compilation compilation, ISymbol symbol, CancellationToken cancellationToken);
     }
 }
